Added name of application to the manifest file. added summary to the daily Gson data class. Changed the return type of longitude and latitude in WeatherResults Gson data class.

+++ app/src/main/java/com/dushanesmith/weather/data/weather/WeatherResults.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName
 
 data class WeatherResults (
 
-  @SerializedName("lat"             ) var lat            : Int?                = null,
-  @SerializedName("lon"             ) var lon            : Int?                = null,
+  @SerializedName("lat"             ) var lat            : Double?                = null,
+  @SerializedName("lon"             ) var lon            : Double?                = null,
   @SerializedName("timezone"        ) var timezone       : String?             = null,
   @SerializedName("timezone_offset" ) var timezoneOffset : Int?                = null,
   @SerializedName("current"         ) var current        : Current?            = Current(),
